+++ src/formatter/usage_formatter_spec.js
@@ -23,56 +23,12 @@ describe('UsageFormatter', function() {
     })
 
     describe('no step definitions', function() {
-      describe('without steps', function() {
-        beforeEach(function() {
-          this.eventBroadcaster.emit('test-run-finished')
-        })
-
-        it('outputs "No step definitions"', function() {
-          expect(this.output).to.eql('No step definitions')
-        })
+      beforeEach(function() {
+        this.eventBroadcaster.emit('test-run-finished')
       })
 
-      describe('with a step', function() {
-        beforeEach(function() {
-          const events = Gherkin.generateEvents(
-            'Feature: a\nScenario: b\nWhen abc\nThen ab',
-            'a.feature'
-          )
-          events.forEach(event => {
-            this.eventBroadcaster.emit(event.type, event)
-            if (event.type === 'pickle') {
-              this.eventBroadcaster.emit('pickle-accepted', {
-                type: 'pickle-accepted',
-                pickle: event.pickle,
-                uri: event.uri
-              })
-            }
-          })
-          const testCase = { sourceLocation: { uri: 'a.feature', line: 2 } }
-          this.eventBroadcaster.emit('test-case-prepared', {
-            ...testCase,
-            steps: [
-              { sourceLocation: { uri: 'a.feature', line: 3 } },
-              { sourceLocation: { uri: 'a.feature', line: 4 } }
-            ]
-          })
-          this.eventBroadcaster.emit('test-step-finished', {
-            index: 0,
-            testCase,
-            result: {}
-          })
-          this.eventBroadcaster.emit('test-step-finished', {
-            index: 1,
-            testCase,
-            result: {}
-          })
-          this.eventBroadcaster.emit('test-run-finished')
-        })
-
-        it('outputs "No step definitions"', function() {
-          expect(this.output).to.eql('No step definitions')
-        })
+      it('outputs "No step definitions"', function() {
+        expect(this.output).to.eql('No step definitions')
       })
     })
